Rename newparser to Parsely
We're going to actually use this so let's give it a more permanent name.
